+++ go/libraries/doltcore/merge/row_merge_test.go
@@ -219,7 +219,7 @@ func TestRowMerge(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			v := newValueMerger(test.mergedSch, test.leftSch, test.rightSch, test.baseSch)
+			v := newValueMerger(test.mergedSch, test.leftSch, test.rightSch, test.baseSch, syncPool)
 
 			merged, isConflict := v.tryMerge(test.row, test.mergeRow, test.ancRow)
 			assert.Equal(t, test.expectConflict, isConflict)
