+++ go/store/prolly/map.go
@@ -165,6 +165,11 @@ func (m Map) IterRange(ctx context.Context, rng Range) (MapIter, error) {
 	}
 }
 
+// Pool returns the pool.BuffPool of the underlying tuples' tree.NodeStore
+func (m Map) Pool() pool.BuffPool {
+	return m.tuples.ns.Pool()
+}
+
 func (m Map) pointLookupFromRange(ctx context.Context, rng Range) (*pointLookup, error) {
 	search := pointLookupSearchFn(rng)
 	cur, err := tree.NewCursorFromSearchFn(ctx, m.tuples.ns, m.tuples.root, search)
